% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_master_csv}
\alias{get_master_csv}
\title{Function to get the latest symbol of exchanges.}
\usage{
get_master_csv(object, param)
}
\arguments{
\item{object}{An object of type espresso connect with valid api_key and
access_token.}

\item{param}{This contains the request parameters.}
}
\value{
Returns a json object with exchnage symbol details, if successful.
}
\description{
Gets all the latest symbols of all the exchanges from the symbol
master files as CSV without access-token authorization.
}
