% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{set_access_token}
\alias{set_access_token}
\title{Function to set the access token in the espresso connect object.}
\usage{
set_access_token(object, access_token)
}
\arguments{
\item{object}{a espresso connect object.}

\item{access_token}{Access token.}
}
\value{
espresso connect object with access token set.
}
\description{
This function sets the access token (obtained by sending a
hash of api key, request token and api secret) inside the espresso connect
object.
}
\seealso{
\code{\link[espressoconnect]{create_connection_object}}

\code{\link[espressoconnect]{fetch_access_token}}
}
