% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{set_request_token}
\alias{set_request_token}
\title{Function to set the request token in the espresso connect object.}
\usage{
set_request_token(object, request_token)
}
\arguments{
\item{object}{an S4 object of type espressoconnect.}

\item{request_token}{a string containing the request token from login flow.}
}
\value{
Returns the object with request token field set.
}
\description{
Function to set request token (obtained from a successful
login flow) in the espresso connect object.
}
\details{
This function allows to set the request token of the espresso
connect object. The request token is obtained from a successful login flow.
After the login flow, the user is redirected to a pre-specified URL. The
request token is posted alongwith as a query parameter.
}
\seealso{
\code{\link[espressoconnect]{create_connection_object}}
}
